remove deprecated Deno APIs
