+++ cli/tests/unit/files_test.ts
@@ -1,163 +1,16 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-// deno-lint-ignore-file no-deprecated-deno-api
-
 import {
   assert,
   assertEquals,
   assertRejects,
   assertThrows,
 } from "./test_util.ts";
-import { copy } from "../../../test_util/std/io/util.ts";
 
 Deno.test(function filesStdioFileDescriptors() {
   assertEquals(Deno.stdin.rid, 0);
   assertEquals(Deno.stdout.rid, 1);
   assertEquals(Deno.stderr.rid, 2);
-});
-
-Deno.test({ permissions: { read: true } }, async function filesCopyToStdout() {
-  const filename = "cli/tests/testdata/fixture.json";
-  const file = await Deno.open(filename);
-  assert(file.rid > 2);
-  const bytesWritten = await copy(file, Deno.stdout);
-  const fileSize = Deno.statSync(filename).size;
-  assertEquals(bytesWritten, fileSize);
-  file.close();
-});
-
-Deno.test({ permissions: { read: true } }, async function filesIter() {
-  const filename = "cli/tests/testdata/hello.txt";
-  const file = await Deno.open(filename);
-
-  let totalSize = 0;
-  for await (const buf of Deno.iter(file)) {
-    totalSize += buf.byteLength;
-  }
-
-  assertEquals(totalSize, 12);
-  file.close();
-});
-
-Deno.test(
-  { permissions: { read: true } },
-  async function filesIterCustomBufSize() {
-    const filename = "cli/tests/testdata/hello.txt";
-    const file = await Deno.open(filename);
-
-    let totalSize = 0;
-    let iterations = 0;
-    for await (const buf of Deno.iter(file, { bufSize: 6 })) {
-      totalSize += buf.byteLength;
-      iterations += 1;
-    }
-
-    assertEquals(totalSize, 12);
-    assertEquals(iterations, 2);
-    file.close();
-  },
-);
-
-Deno.test({ permissions: { read: true } }, function filesIterSync() {
-  const filename = "cli/tests/testdata/hello.txt";
-  const file = Deno.openSync(filename);
-
-  let totalSize = 0;
-  for (const buf of Deno.iterSync(file)) {
-    totalSize += buf.byteLength;
-  }
-
-  assertEquals(totalSize, 12);
-  file.close();
-});
-
-Deno.test(
-  { permissions: { read: true } },
-  function filesIterSyncCustomBufSize() {
-    const filename = "cli/tests/testdata/hello.txt";
-    const file = Deno.openSync(filename);
-
-    let totalSize = 0;
-    let iterations = 0;
-    for (const buf of Deno.iterSync(file, { bufSize: 6 })) {
-      totalSize += buf.byteLength;
-      iterations += 1;
-    }
-
-    assertEquals(totalSize, 12);
-    assertEquals(iterations, 2);
-    file.close();
-  },
-);
-
-Deno.test(async function readerIter() {
-  // ref: https://github.com/denoland/deno/issues/2330
-  const encoder = new TextEncoder();
-
-  class TestReader implements Deno.Reader {
-    #offset = 0;
-    #buf: Uint8Array;
-
-    constructor(s: string) {
-      this.#buf = new Uint8Array(encoder.encode(s));
-    }
-
-    read(p: Uint8Array): Promise<number | null> {
-      const n = Math.min(p.byteLength, this.#buf.byteLength - this.#offset);
-      p.set(this.#buf.slice(this.#offset, this.#offset + n));
-      this.#offset += n;
-
-      if (n === 0) {
-        return Promise.resolve(null);
-      }
-
-      return Promise.resolve(n);
-    }
-  }
-
-  const reader = new TestReader("hello world!");
-
-  let totalSize = 0;
-  for await (const buf of Deno.iter(reader)) {
-    totalSize += buf.byteLength;
-  }
-
-  assertEquals(totalSize, 12);
-});
-
-Deno.test(async function readerIterSync() {
-  // ref: https://github.com/denoland/deno/issues/2330
-  const encoder = new TextEncoder();
-
-  class TestReader implements Deno.ReaderSync {
-    #offset = 0;
-    #buf: Uint8Array;
-
-    constructor(s: string) {
-      this.#buf = new Uint8Array(encoder.encode(s));
-    }
-
-    readSync(p: Uint8Array): number | null {
-      const n = Math.min(p.byteLength, this.#buf.byteLength - this.#offset);
-      p.set(this.#buf.slice(this.#offset, this.#offset + n));
-      this.#offset += n;
-
-      if (n === 0) {
-        return null;
-      }
-
-      return n;
-    }
-  }
-
-  const reader = new TestReader("hello world!");
-
-  let totalSize = 0;
-  for await (const buf of Deno.iterSync(reader)) {
-    totalSize += buf.byteLength;
-  }
-
-  assertEquals(totalSize, 12);
 });
 
 Deno.test(
