+++ runtime/js/90_deno_ns.js
@@ -63,14 +63,6 @@
     env: __bootstrap.os.env,
     exit: __bootstrap.os.exit,
     execPath: __bootstrap.os.execPath,
-    Buffer: __bootstrap.buffer.Buffer,
-    readAll: __bootstrap.buffer.readAll,
-    readAllSync: __bootstrap.buffer.readAllSync,
-    writeAll: __bootstrap.buffer.writeAll,
-    writeAllSync: __bootstrap.buffer.writeAllSync,
-    copy: __bootstrap.io.copy,
-    iter: __bootstrap.io.iter,
-    iterSync: __bootstrap.io.iterSync,
     SeekMode: __bootstrap.io.SeekMode,
     read: __bootstrap.io.read,
     readSync: __bootstrap.io.readSync,
